pyproject.toml
[project]
name = "hackathon-0"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
